+++ Tools/ToolSumBrownianMotions/parameters.h
@@ -16,7 +16,7 @@ struct Parameters
 
   }
   Parameters(
-    const double any_n,
+    const int any_n,
     const double any_mean_a,
     const double any_stddev_a,
     const double any_mean_b,
